FS-4082: add pagination logic for sections on an application
On the application detail page (task list into individual sections),
allow moving between section detail pages with a simple next/ previous
pagination component.

We use the govuk design system `pagination` component which is supported
by our pinned version of the design system and the jinja component
library used.

The `state` dictionary which is consumed by every detail route describes
the application and the assessors progress through it, as it already
loaded information on all of the sections this seemed like a good place
to add a utility for pagination. This utility can then also be used
anywhere future tickets may need to move forwards and backwards between
sections.

This learns from the functionality but deviates from the implementation
of a work in progress branch in a number of ways
https://github.com/communitiesuk/funding-service-design-assessment/tree/FS-4082-assessments-list--prev-next-links-v1

- identifies that the information on sections already seems like it has
  been fetched to load the page (through the `state` abstractin), by
  adding an accessor to this existing class we should benefit
    - four less HTTP calls, some of which are dependant on the size of the
      application
    - avoid fetching application details where they won't be used on the
      page (potentially easy to leak into logs or the template if future
      devs aren't aware of this)
    - configure the business logic somewhere that should be easier to
      cover with smaller tests
    - allow the forward/ backwards logic to be used anywhere that
      sections are mentioned in the future
- takes the next and previous logic out of the template to make it
  easier to cover
- removes any query strings used for logic, the aim is to make loading
  the pages independant of any outside state and simplify the URLs that
  users might be passing around to each other
- uses the design system pagination component for accessibility and
  maintainability (this should track upstream changes from the design
  system team)