FS-4082: unit coverage for different pagination scenarios
This coverage focusses on the behaviour of sub-sections across groups
that can be defined in either the `sections` or `criteria` schema. The
task list component treats these the same and the behaviour described
should allow the user to go "next" and "previous" between these groups.

We check that that the parser and behaviour work for the shared app
fixture used to test the route controller behaviour.

We also check for exceptional cases using simpler stubs.