FS-4082: move common pagination markup into a component snippet
The sub section pagination is shared by the `scores` and `assessments`
blueprints. As it relies on the a consistent helper model move it into a
component that can be included by those separate templates.

This should keep any changes or updates consistent between the pages
(that are directly linked from the users point of view).