FS-4674: visually indicate funds that with the "EOI" funding type
Now that funds have the `funding_type` property add a visual indiciation
any time that we show the "noun" fund "name" property on the dashboard pages.

Note this isn't the same as the "titles" properties which usually are
verbs and translated.

Future fund models may model the EOI as a proprety on the fund but for now the
questions and configuration for expression of interests are effectively
modelled as funds themselves.