FS-4647: visually differentiate expresssion of interest funds
This adds a suffix to the name of funds with the `funding_type` property "EOI".

Note that the fund name is used in many places in the assessment tool so for now
this just sets the suffix in the data model, ideally this should be a
presentation layer (template) concern and be done by jinja based on the funds properties.
This however does the job.

Commiting for opinions and consensus.