Increase CPU for pre-award-stores
This app will be doing the work of three stores soon, and giving it 0.5
CPU is leading to significant throttling/delays when processing even a
small amount of concurrent users (e2e tests) in our lower environments.

Let's give it a (barely) more respectable resource allocation.

This will be something to continue to monitor/tweak as we bring in more
apps, eg frontends.