Address ruff linting error
B008 Do not perform function call `get_ttl_hash` in argument defaults;
instead, perform the call within the function, or read the default from
a module-level singleton variable

It looks like all of the caching calls we do are a little bit funny - we
pass in a parameter that is meant to cache bust after some amount of
time passes and then manually `del` that parameter. This feels unusual.
I think we could just use TTLCache. But leaving for the future...