Pull in language selector from utils and fix for host matching
Now that all of our frontends are coming in to one place, we don't need
to use this from utils any more.

This pulls it into the pre-award-frontend app, and changes
LanguageSelector to register the route against all hosts.

This fixes a bug where the `url_for` on `select_language` was resolving
to an invalid URL, preventing users from changing language to Welsh on
eg COF.