Ignore OS/permission errors writing this file
When running DB migrations in our paketo containers, the user does not
have permission to write to the filesystem in the code directory, so it
throws a PermissionError.

We can just ignore this here. This code is intended to be run/used
locally by developers to help stop two PRs having migrations with the
same down revision.