FSPT-189: Switch session cookie to `Lax` for Talisman
Authenticator used a session cookie with Lax rather than Strict for the
`session_cookie_samesite` in Talisman so that it could send people to
Microsoft logins. Assess had it set to Strict for this same setting.

Now that we've brought the frontends together this needs to be set to
Lax otherwise you can't log into Find or Submit.