Remove connexion
We switch from gunicorn with uvicorn workers (WSGI server, ASGI worker
that plugs into connexion, which is an ASGI framework) to gunicorn with
gevent workers (WSGI server, WSGI worker, with flask which is a WSGI
framework).

The `gevent` workers use green threads/greenlets to achieve cooperative
concurrency. When using these workers, we need to "monkeypatch" all of
the IO-bound call sites as early in the python process as possible. When
running with gunicorn, the first file that gets sourced is the gunicorn
config, so we monkeypatch the IO calls there (eg ssl, socket,
filesystem, etc)