Run with 3 gunicorn workers
This is what pre-award-frontend does, and should allow us to make
slightly more full use of our deployed apps.

However - we should probably be detecting this dynamically based on the
number of cores available to us, as the recommendation is generally to
start with (2 * cores) + 1 workers.

For now our deployed instances have 1 CPU, so 3 is about right.