Opt in to python e2e tests for dev+test environments
Switched from opt-out to opt-in for these tests, because we need to not
run them in the UAT environment.

The logic here looks slightly broken; I think that false values would
end up being set to true because of `||`. But that's pre-existing ...
and deployments are currently broken, so prioritising that. (Also we're
going to revisit pipelines fairly wholesale soon - ✨  consolidation ✨
)