Fix translations
This needed to be updated when changing from Flask-Babel v2 to
Flask-Babel v4, but I missed it. I missed it because half of our
translations were working and half were not. Because, I think, some
strings are translated in python and some are translated in jinja.
The python ones were working, but the jinja ones seemed to not be.

From the changelog:

- Babel.locale_selector and Babel.timezone_selector no longer exist.
They must be provided either when the Babel() object is created or
when init_app() is called. This is to support having a single Babel
object for multiple Flask apps (#107) as well as to simplify
settings
and multi-threaded state.