use app-specific api hosts
We are making API calls over service connect rather than public DNS.
This will talk to nginx for basic auth in dev/test/uat, and direct to
flask in production.

We need new secrets so that existing pre-award-frontend can talk to
existing pre-award-stores, but the new integrated pre-award can talk
over service connect addresses.

The basic auth username and password need to be baked into the URLs in
dev/test/uat. The port in those envs will also be the nginx sidecar, not
the main app container.

This does make production a bit of a snowflake case. Hopefully it
doesn't break..?

FOR SOME REASON, API_HOST should *not* include the port :think: I do not
understand this yet because I definitely thought the port was part of
the host for URL routing. But in deployed environments (AWS), the port
is for some reason not present in `request.host` when it reaches the
app. Maybe ECS is doing some magic ... or maybe I am just
misunderstanding/misremembering some of the flask stuff around host
matching/routing. One to explore/investigate a bit further...