Add API_HOST for production
In dev/test/uat, we run an nginx sidecar that changes to ingress port
from the standard 8080 to 8087. The sidecar is used to add basic auth to
those non-prod environments.

In prod, the sidecar doesn't exist and so it's just available on the
native 8080 port.

Importantly, the nginx sidecar also strips the port out of the Host
header, so in non-prod environments API_HOST needs to *not* include the
port, and in prod it must have the port.