FSPT-192: very temporay app source metadata
One method of including information about which app a request was for in the logs.

We expect that the stores/ frontend boundaries will be removed very soon
so this doesn't need to exist for very long, in the meantime it should
help developers see which context a log came from.

In the future the top level blueprints that are decided on could specify
their own `@bp.after_request` which could include a common key
representing the domain boundary (likely "apply" or "assess").

As the "API_HOST" is shared between the 4 stores (application,
assessment, account and fund) log the string version of the blueprint
used to serve that API request - this relies on the convention that the
blueprints are named appropriately but should work for now.