BAU: fix host matching for authenticator errors
`AUTHENTICATOR_HOST` looks like it refers to the publicly addressable
URL for authenticator (used to build links), `AUTH_HOST` is the internal
server name we'd expect to be passed in the request header.

This confusion should only exist in the short period the apps and stores
are combined but haven't had the network layer removed/ been
de-duplicated.