Address feedback
Addressing the feedback.

There are few important changes:
- get_flags is now only returning non change requests
- added get_change_requests function which only returns change requests

Changed so the set the AssessmentRecord.workflow_status to CHANGE_REQUESTED
on POST change_request.

There is no endpoint on assessment_store to update the flag
so I had to make use of the model directly inside the data_services.py.

Determining the status of the Application no longer relies on
Change Requests flags, directly uses AssessmentRecord.workflow_status.

Flags still override the final status of the AssessmentRecord if they exist.